update stylelint rule

+++ .stylelintrc.js
@@ -39,7 +39,7 @@ module.exports = {
         ],
 
         // Opinionated rule, you can disable it if you want
-        "string-quotes": "single",
+        "string-quotes": "double",
 
         // https://github.com/hudochenkov/stylelint-order/blob/master/rules/order/README.md
         "order/order": ["custom-properties", "dollar-variables", "declarations", "at-rules", "rules"],
